Pretty huge update.
We have super basic switching of 2 elements for debugging, some visual code for auras, placement of Earth projectiles and quite a few fixes/adjustments to some animations. The walk cycle still looks awful despite me attempting to draw hammerpants/hakama to make the walk cycle look less weird. We very importantly have the beginning of a charge mechanic, where our character gets to both move faster and get access to stronger attacks if they don't fire.

Next up is probably a big clean up of our character for demonstration purposes as a lot of his sprites are turned off. It'll be one last push to be the character look semi-presentable, and to demonstrate how we want the flow of gameplay to work. Maybe we'll eventually get a good looking walk cycle.

Also, we really need to let the character move properly and collide correctly, meaning we're staring into the abyss of trying to learn raycast movement again.